Improve cart-in-iframe compatibility
These people use the exact same stuff both on the front-end and on the back office. This creates a whole bunch of issues related to colors and contrast. And of course no way to target elements with css. If I could I would happily kill them. I hate them. Seriously.

Signed-off-by: Emanuele <feliziani.emanuele@gmail.com>