Gzip css and js in the build process
Update che CodeKit config file to include a build process that gzips css and js files so that they can be uploaded to S3 already zipped.

The config file contains a whole lot of other stuff that CodeKit is keeping track of. Most of it is none of my business.

The commit also ignores the gzipped files in the repository. We have the source, we don't need the output in here.

Signed-off-by: Emanuele <feliziani.emanuele@gmail.com>